add UIView extension function to support auto layout in this pod change version number to 2.0.7 in .podspec
