+++ HushPhotoPicker.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
   s.name             = 'HushPhotoPicker'
-  s.version          = '2.0.6'
+  s.version          = '2.0.7'
   s.summary          = 'multiple phassets picker for iOS lib. like facebook'
 
 # This description is used to generate tags and improve search results.
